+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 `utopya` aims to adhere to [semantic versioning](https://semver.org/).
 
+## v1.2.2
+- !46 fixes a bug in the ABM plot occurring with the latest xarray version.
+
+
 ## v1.2.1
 - !43 allows model-specific updates to the `Multiverse` configuration
 - !43 makes specifying a model executable optional, allowing *evaluation only* pipelines.